// src/enums/PositionUnit.ts
/**
 * Position unit enumeration
 * Defines the different types of position units
 */
export enum PositionUnit {
  PIXEL = 'pixel',
  VIEWPORT = 'viewport',
  PERCENT = 'percent',
  PERCENTAGE = 'percentage',
  PARENT_WIDTH = 'parent-width',
  PARENT_HEIGHT = 'parent-height',
  PARENT_LEFT = 'parent-left',
  PARENT_TOP = 'parent-top',
  PARENT_CENTER_X = 'parent-center-x',
  PARENT_CENTER_Y = 'parent-center-y',
  VIEWPORT_WIDTH = 'viewport-width',
  VIEWPORT_HEIGHT = 'viewport-height',
  VIEWPORT_LEFT = 'viewport-left',
  VIEWPORT_TOP = 'viewport-top',
  VIEWPORT_CENTER_X = 'viewport-center-x',
  VIEWPORT_CENTER_Y = 'viewport-center-y',
  SCENE_CENTER_X = 'scene-center-x',
  SCENE_CENTER_Y = 'scene-center-y',
  CONTENT_LEFT = 'content-left',
  CONTENT_RIGHT = 'content-right',
  CONTENT_TOP = 'content-top',
  CONTENT_BOTTOM = 'content-bottom',
  CENTER = 'center',
  LEFT = 'left',
  RIGHT = 'right',
  TOP = 'top',
  BOTTOM = 'bottom',
  RANDOM = 'random'
}